requirements.txt
Flask
torch
torchvision
timm
Pillow
